backend/app/fastapi_verse_search/requirements.txt
fastapi
uvicorn
faiss-cpu
torch
sentence-transformers
numpy
scikit-learn
python-dotenv
datasets
tqdm
annoy